fix: keep procedural card stroke width typed as double

+++ mobile/lib/widgets/zync_card_preview.dart
@@ -282,7 +282,7 @@ class _CardScenePainter extends CustomPainter {
     final line = Paint()
       ..color = Colors.white.withValues(alpha: 0.54)
       ..style = PaintingStyle.stroke
-      ..strokeWidth = math.max(2, size.width * 0.012);
+      ..strokeWidth = math.max(2.0, size.width * 0.012).toDouble();
 
     final court = RRect.fromRectAndRadius(
       Rect.fromLTWH(
